added memory address struct, updated ui.

+++ Modern Warfare 2 - FOV Changer/src/MemoryManager.cpp
@@ -17,7 +17,7 @@ MemoryManager::~MemoryManager()
 		CloseHandle(hProc);
 }
 
-DWORD MemoryManager::getProcID() const
+unsigned long MemoryManager::getProcID() const
 {
 	return procID;
 }
